Update Ray Tracing Scripts and Add New Features
Updated ray tracing scripts to apply multiple enhancements including the addition of energy, light bounce mode, and sun colour. Also introduced the ability to calculate depth and brightness for a specific hit. Added denoiser and ray tracing animator scripts. Integrated functionality to open screenshots and minor adjustments in the unity file. Begun working on an animator (RaytracingAnimator.cs) that will move the camera along a set path when a frame is fully rendered and denoised to allow animations/cinematic videos to be created.